+++ src/main/kotlin/app/revanced/patches/reddit/customclients/boostforreddit/api/patch/ChangeOAuthClientIdPatch.kt
@@ -1,34 +1,29 @@
 package app.revanced.patches.reddit.customclients.boostforreddit.api.patch
 
-import app.revanced.extensions.toErrorResult
-import app.revanced.patcher.annotation.*
+import app.revanced.patcher.annotation.Compatibility
+import app.revanced.patcher.annotation.Package
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprintResult
 import app.revanced.patcher.patch.PatchResult
 import app.revanced.patcher.patch.PatchResultSuccess
-import app.revanced.patcher.patch.annotations.Patch
 import app.revanced.patches.reddit.customclients.AbstractChangeOAuthClientIdPatch
+import app.revanced.patches.reddit.customclients.ChangeOAuthClientIdPatchAnnotation
 import app.revanced.patches.reddit.customclients.boostforreddit.api.fingerprints.GetClientIdFingerprint
 
-@Patch
-@Name("change-oauth-client-id")
-@Description("Changes the OAuth client ID in Boost for Reddit.")
+@ChangeOAuthClientIdPatchAnnotation
 @Compatibility([Package("com.rubenmayayo.reddit")])
-@Version("0.0.1")
 class ChangeOAuthClientIdPatch : AbstractChangeOAuthClientIdPatch(
-    "http://rubenmayayo.com",
-    Options,
-    GetClientIdFingerprint
+    "http://rubenmayayo.com", Options, listOf(GetClientIdFingerprint)
 ) {
-    override fun MethodFingerprint.patch(context: BytecodeContext): PatchResult {
-        result?.mutableMethod?.addInstructions(
+    override fun List<MethodFingerprintResult>.patch(context: BytecodeContext): PatchResult {
+        first().mutableMethod.addInstructions(
             0,
-            """
+    """
                      const-string v0, "$clientId"
                      return-object v0
                 """
-        ) ?: return toErrorResult()
+        )
 
         return PatchResultSuccess()
     }