+++ src/main/kotlin/app/revanced/patches/youtube/layout/autocaptions/fingerprints/SubtitleButtonControllerFingerprint.kt
@@ -6,7 +6,10 @@ import org.jf.dexlib2.AccessFlags
 import org.jf.dexlib2.Opcode
 
 object SubtitleButtonControllerFingerprint : MethodFingerprint(
-    "V", AccessFlags.PUBLIC or AccessFlags.FINAL, listOf("L"), listOf(
+    returnType = "V",
+    accessFlags = AccessFlags.PUBLIC or AccessFlags.FINAL,
+    parameters = listOf("Lcom/google/android/libraries/youtube/player/subtitles/model/SubtitleTrack;"),
+    opcodes = listOf(
         Opcode.IGET_OBJECT,
         Opcode.IF_NEZ,
         Opcode.RETURN_VOID,
@@ -16,5 +19,8 @@ object SubtitleButtonControllerFingerprint : MethodFingerprint(
         Opcode.CONST,
         Opcode.INVOKE_VIRTUAL,
         Opcode.IGET_OBJECT,
-    )
+    ),
+    customFingerprint = { methodDef, _ ->
+        methodDef.definingClass.endsWith("SubtitleButtonController;")
+    }
 )