+++ src/main/kotlin/app/revanced/patches/youtube/layout/hide/infocards/fingerprints/InfocardsIncognitoFingerprint.kt
@@ -5,7 +5,8 @@ import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
 import org.jf.dexlib2.AccessFlags
 
 object InfocardsIncognitoFingerprint : MethodFingerprint(
-    "Ljava/lang/Boolean;",
-    AccessFlags.PUBLIC or AccessFlags.FINAL,
+    accessFlags = AccessFlags.PUBLIC or AccessFlags.FINAL,
+    returnType = "Ljava/lang/Boolean;",
+    parameters = listOf("L", "J"),
     strings = listOf("vibrator")
 )