+++ src/main/kotlin/app/revanced/patches/youtube/layout/hide/shorts/bytecode/fingerprints/CreateShortsButtonsFingerprint.kt
@@ -1,19 +1,13 @@
 package app.revanced.patches.youtube.layout.hide.shorts.bytecode.fingerprints
 
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import app.revanced.patcher.extensions.or
+import app.revanced.util.patch.LiteralValueFingerprint
 import app.revanced.patches.youtube.layout.hide.shorts.resource.patch.HideShortsComponentsResourcePatch
-import org.jf.dexlib2.Opcode
-import org.jf.dexlib2.iface.instruction.WideLiteralInstruction
+import org.jf.dexlib2.AccessFlags
 
-object CreateShortsButtonsFingerprint : MethodFingerprint(
+object CreateShortsButtonsFingerprint : LiteralValueFingerprint(
+    accessFlags = AccessFlags.PRIVATE or AccessFlags.FINAL,
+    returnType = "V",
     parameters = listOf("Z", "Z", "L"),
-    customFingerprint = { methodDef, _ ->
-        methodDef.implementation?.instructions?.any {
-            if (it.opcode != Opcode.CONST) return@any false
-
-            val literal = (it as WideLiteralInstruction).wideLiteral
-
-            literal == HideShortsComponentsResourcePatch.reelPlayerRightLargeIconSize
-        } ?: false
-    }
+    literal = HideShortsComponentsResourcePatch.reelPlayerRightLargeIconSize
 )