+++ src/main/kotlin/app/revanced/patches/youtube/video/oldqualitylayout/fingerprints/QualityMenuViewInflateFingerprint.kt
@@ -1,11 +1,15 @@
 package app.revanced.patches.youtube.video.oldqualitylayout.fingerprints
 
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import app.revanced.patcher.extensions.or
+import app.revanced.util.patch.LiteralValueFingerprint
 import app.revanced.patches.youtube.video.oldqualitylayout.patch.OldQualityLayoutResourcePatch
+import org.jf.dexlib2.AccessFlags
 import org.jf.dexlib2.Opcode
-import org.jf.dexlib2.iface.instruction.WideLiteralInstruction
 
-object QualityMenuViewInflateFingerprint : MethodFingerprint(
+object QualityMenuViewInflateFingerprint : LiteralValueFingerprint(
+    accessFlags = AccessFlags.PUBLIC or AccessFlags.FINAL,
+    parameters = listOf("L", "L", "L"),
+    returnType = "L",
     opcodes = listOf(
         Opcode.INVOKE_SUPER,
         Opcode.CONST,
@@ -22,13 +26,5 @@ object QualityMenuViewInflateFingerprint : MethodFingerprint(
         Opcode.MOVE_RESULT_OBJECT,
         Opcode.CHECK_CAST
     ),
-    customFingerprint = { methodDef, _ ->
-        methodDef.implementation?.instructions?.any {
-            if (it.opcode != Opcode.CONST) return@any false
-
-            val literal = (it as WideLiteralInstruction).wideLiteral
-
-            literal == OldQualityLayoutResourcePatch.videoQualityBottomSheetListFragmentTitle
-        } ?: false
-    }
+    literal = OldQualityLayoutResourcePatch.videoQualityBottomSheetListFragmentTitle
 )